+++ languages/de/tl_anystores_settings.php
@@ -3,7 +3,7 @@
 /**
  * anyStores for Contao Open Source CMS
  *
- * @copyright   (c) 2015 Tastaturberuf <mail@tastaturberuf.de>
+ * @copyright   (c) 2015, 2016 Tastaturberuf <mail@tastaturberuf.de>
  * @author      Daniel Jahnsmüller <mail@jahnsmueller.net>
  * @license     http://opensource.org/licenses/lgpl-3.0.html
  * @package     anyStores
@@ -12,10 +12,14 @@
 
 $GLOBALS['TL_LANG']['tl_anystores_settings'] = array
 (
+
     // Legends
-    'anystores_api_legend' => 'API-Einstellungen',
+    'anystores_common_legend' => 'Allgemeine Einstellungen',
+    'anystores_api_legend'    => 'API-Einstellungen',
 
     // Fields
-    'anystores_geoApi'        => array('Auswahl des API-Anbieters', 'Wählen Sie einen API-Anbieter aus der Liste.'),
-    'anystores_defaultMarker' => array('Standard-Marker für alle Standorte', 'Kann im Modul, in der Kategorie und am Standort überschrieben werden.')
+    'anystores_defaultMarker' => array('Standard-Marker für alle Standorte', 'Kann im Modul, in der Kategorie und am Standort überschrieben werden.'),
+    'anystores_geoApi'        => array('API-Anbieter', 'Wählen Sie einen API-Anbieter aus.'),
+    'anystores_apiKey'        => array('API-Key', 'Geben Sie Ihren API-Key ein.')
+
 );