+++ languages/en/tl_anystores_settings.php
@@ -3,7 +3,7 @@
 /**
  * anyStores for Contao Open Source CMS
  *
- * @copyright   (c) 2015 Tastaturberuf <mail@tastaturberuf.de>
+ * @copyright   (c) 2015, 2016 Tastaturberuf <mail@tastaturberuf.de>
  * @author      Daniel Jahnsmüller <mail@jahnsmueller.net>
  * @license     http://opensource.org/licenses/lgpl-3.0.html
  * @package     anyStores
@@ -12,10 +12,14 @@
 
 $GLOBALS['TL_LANG']['tl_anystores_settings'] = array
 (
+
     // Legends
-    'anystores_api_legend' => 'API Settings',
+    'anystores_common_legend' => 'Common settings',
+    'anystores_api_legend'    => 'API settings',
 
     // Fields
+    'anystores_defaultMarker' => array('Default Marker for all locations', 'Can be overriden in module, in categoy or in the location.'),
     'anystores_geoApi'        => array('API provider', 'Choose an API provider from the list above.'),
-    'anystores_defaultMarker' => array('Default Marker for all locations', 'Can be overriden in module, in categoy or in the location.')
+    'anystores_apiKey'        => array('API key', 'Put your API key.')
+
 );